Project3/agents/fetcher/requirements.txt
yfinance
requests
feedparser
python-dotenv
